refactor(scripts): Extract shared _gh helper, drop preflight allocation
- `resolve_repo()` and `gh_repo()` were duplicated between
  `scripts/pull_reviews.py` and `scripts/reply_review.py`. Extract
  them into a private sibling module `scripts/_gh.py` so future
  updates land in one place. Both CLIs invoke it via `from _gh
  import resolve_repo` — direct-script invocation puts `scripts/`
  on `sys.path[0]` automatically, so no package plumbing is needed.
- The PR-existence preflight used `subprocess.check_output` and
  discarded the full PR JSON response. Switch to
  `subprocess.run(..., stdout=subprocess.DEVNULL, check=True)` so
  the body isn't captured into memory on every invocation.
- Add `__pycache__/` to .gitignore since Python now writes bytecode
  for `_gh` into `scripts/__pycache__/` on first import.
