+++ .gitignore
@@ -2,6 +2,7 @@
 debug/
 target/
 Cargo.lock
+__pycache__/
 
 # Editor
 .vscode/
